[perl #31659] [PATCH] Parrot m4 0.0.8
this revision of Parrot m4 has no new features.
Sorry, it isn't ready for 'autoconf' yet :;).

I fiddled a little bit with the harness script.

Mostly I played with the 'eval' command, which is actually a compiler
implemented in C.
The generated bytecode should now be more conformant to the Parrot calling
conventions.
The problems with the 'eval' tests were propable due to a missing 'end' in
the generated bytecode.

Courtesy of Bernhard Schmalhofer <Bernhard.Schmalhofer@biomax.de>


git-svn-id: http://svn.perl.org/parrot/trunk@6682 d31e2699-5ff4-0310-a27c-f18f2fbe73fe